refactor: change endpoint to search and change http method to get
